Dev

See merge request wallet/client/wallet_sdk!88

+++ wallet-database/src/entities/api_nonce.rs
@@ -5,7 +5,7 @@ pub struct ApiNonceEntity {
     pub name: String,
     pub from_addr: String,
     pub chain_code: String,
-    pub nonce: i32,
+    pub nonce: u64,
     pub created_at: sqlx::types::chrono::DateTime<sqlx::types::chrono::Utc>,
     pub updated_at: Option<sqlx::types::chrono::DateTime<sqlx::types::chrono::Utc>>,
 }
